Create RoboTalk.java
I have created  the general skeleton structure for the Robot talk program. I still need to find some modules to perfor  the functions listed . Anyone who has a good understanding in Java can contribute to this repo or are willing to put in the time to add some sort of contribution, Even commenting the code is greatly appreciated . To contribute, just request me so i can give you privelages to write on the RoboTalk repo or you can just clone the repo since it is public anyway . .Regardless please let me know if you want to contribute so dont want this repo to be ruined. Cheers :))